Applying research findings to improve the GDN (in GDN.py) leading to better accuracy and speed.
The master's thesis "Anomaly Detection in the EtherCAT Network of a Power Station" conducted by Niklas Barth at KTH Royal Institute of Technology as part of the ICT Innovation Data Science M.Sc. found that the structure of the learned GDN can be improved by applying a similarity thresholding technique when determining relations between nodes. It's been demonstrated that this novel approach leads to superior accuracy and faster computations. Read more in the publicly available thesis in the DiVA portal (available starting in September 2023).